chore(wmg): refactor ColorScale
