+++ frontend/src/views/WheresMyGene/components/Filters/components/ColorScale/types.ts
@@ -1,0 +1,9 @@
+import { InputDropdownProps as IInputDropdownProps } from "@czi-sds/components";
+
+export interface Props {
+  setIsScaled: (prevIsScaled: boolean) => void;
+}
+
+export const DEFAULT_INPUT_DROPDOWN_PROPS: Partial<IInputDropdownProps> = {
+  sdsStyle: "square",
+};
